feat: dnd로 업로드 한 이미지 누끼 딴 후, 캔버스에 그리기
